feat: add modal window and split codebase into modular components
