+++ itprofit/src/style.css
@@ -9,7 +9,7 @@ body {
   width: 100vw;
   height: 130vh;
   background-color: rgba(20, 20, 20, 0.7);
-  z-index: 100;
+  z-index: 105;
 }
 
 .buy-form {
@@ -37,7 +37,7 @@ body {
 .buy-form form .row {
   display: flex;
   flex-wrap: wrap;
-  ow-gap: 15px;
+  row-gap: 15px;
   -moz-column-gap: 15px;
        column-gap: 15px;
 }
@@ -146,7 +146,7 @@ body {
 .error {
   height: 10px;
   color: red;
-  font-size: 12px;
+  font-size: 14px;
   display: block;
   margin-top: 5px;
   text-align: center;
