+++ itprofit/src/style.sass
@@ -1,14 +1,47 @@
 body
   background-color: #3878ac
+  font-family: 'Roboto', sans-serif
 
-.shadow-wrapper
-  position: absolute
-  top: 0
+.modal
+  display: none
+  position: fixed
+  z-index: 1
+  padding-top: 300px
   left: 0
-  width: 100vw
-  height: 130vh
-  background-color: rgba(20, 20, 20, 0.7)
-  z-index: 100
+  top: -100%
+  width: 100%
+  height: 140%
+  overflow: auto
+  background-color: rgb(0, 0, 0)
+  background-color: rgba(0, 0, 0, 0.4)
+  transition: top 1s ease
+
+  &-content
+    background-color: #fefefe
+    margin: auto
+    padding: 20px
+    border: 1px solid #888
+    border-radius: 5px
+    width: 50%
+    height: 15%
+
+  .close
+    color: #aaaaaa
+    float: right
+    font-size: 28px
+    font-weight: bold
+
+    &:hover, &:focus
+      color: #000
+      text-decoration: none
+      cursor: pointer
+
+  #modalText
+    height: 100%
+    display: flex
+    justify-content: center
+    align-items: center
+    font-size: 18px
 
 .buy-form
   display: flex
@@ -19,7 +52,6 @@ body
   margin-top: -300px
   left: 50%
   margin-left: -375px
-  z-index: 102
   display: table-cell
   vertical-align: middle
 
@@ -27,10 +59,12 @@ body
     position: absolute
     padding: 20px
     width: 700px
-    background: #fff
+    background: #dddddd
+    border-radius: 7px
     box-shadow: 0 5px 10px rgba(0, 0, 0, 0.1)
     font-size: 18px
     line-height: 100%
+    margin-top: 10px
 
     .row
       display: flex
@@ -80,17 +114,18 @@ body
   &invalid
     border-color: red
 
-.submit-btn
+.btn
   width: 100%
   padding: 12px
   font-size: 17px
-  background: #27ae60
+  background: #525f61
+  border-radius: 5px
   color: #fff
   margin-top: 5px
   cursor: pointer
 
   &:hover
-    background: #2ecc71
+    background: #627174
 
 @media (max-width: 1100px)
   .buy-form form
@@ -123,20 +158,20 @@ body
 
 @media (max-width: 800px)
   .buy-form
-    margin-top: -380px
+    margin-top: -300px
 
 @media (max-width: 600px)
   .buy-form form
     width: 300px
 
   .buy-form
-    margin-top: -380px
-    margin-left: -150px
+    margin-top: -300px
+    margin-left: -160px
 
 .error
   height: 10px
   color: red
-  font-size: 12px
+  font-size: 14px
   display: block
   margin-top: 5px
   text-align: center
